Delete app-foundations
