+++ api/package.json
@@ -29,7 +29,8 @@
     "build": "tsc && unbuild"
   },
   "dependencies": {
-    "@previewjs/type-analyzer": "^4.0.7"
+    "@previewjs/type-analyzer": "^4.0.7",
+    "axios": "^1.1.3"
   },
   "devDependencies": {
     "unbuild": "0.9.1"
