+++ api/src/index.ts
@@ -1,4 +1,5 @@
 export * from "./api";
+export * from "./axios-api";
 export * from "./component-id";
 export * from "./persisted-state";
 export * from "./rpc";
