--- app-foundations/.eslintrc.js
@@ -1,1 +1,0 @@
-module.exports = require("../.eslintrc.react");
