--- app-foundations/__previewjs__/Wrapper.tsx
@@ -1,4 +1,0 @@
-import { Fragment } from "react";
-import "../../app/client/src/index.css";
-
-export const Wrapper = Fragment;
