--- app-foundations/build.config.ts
@@ -1,9 +1,0 @@
-export default {
-  entries: ["./src/index"],
-  rollup: {
-    emitCJS: true,
-  },
-  externals: ["typescript"],
-  declaration: true,
-  clean: true,
-};
