--- app-foundations/package.json
@@ -1,50 +1,0 @@
-{
-  "name": "@previewjs/app-foundations",
-  "version": "6.0.0",
-  "license": "AGPL-3.0",
-  "author": {
-    "name": "François Wouts",
-    "email": "f@zenc.io"
-  },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/fwouts/previewjs"
-  },
-  "bugs": {
-    "url": "https://github.com/fwouts/previewjs/issues"
-  },
-  "homepage": "https://previewjs.com",
-  "main": "./dist/index.cjs",
-  "module": "./dist/index.mjs",
-  "types": "./dist/index.d.ts",
-  "exports": {
-    ".": {
-      "require": "./dist/index.cjs",
-      "import": "./dist/index.mjs",
-      "types": "./dist/index.d.ts"
-    }
-  },
-  "scripts": {
-    "prepublish": "cd .. && pnpm turbo run build --scope=@previewjs/app-foundations --no-deps --include-dependencies",
-    "build": "tsc && unbuild",
-    "watch": "concurrently 'tsc --watch' 'nodemon -e ts,tsx --watch src -x pnpm unbuild'",
-    "test": "vitest --no-threads"
-  },
-  "dependencies": {
-    "@previewjs/api": "^5.0.0",
-    "@previewjs/iframe": "^3.0.0",
-    "@previewjs/properties": "^1.0.0",
-    "@previewjs/serializable-values": "^3.1.1",
-    "@previewjs/type-analyzer": "^4.0.7",
-    "assert-never": "^1.2.1",
-    "axios": "^1.1.3",
-    "mobx": "^6.6.1"
-  },
-  "devDependencies": {
-    "concurrently": "^7.5.0",
-    "nodemon": "^2.0.20",
-    "typescript": "4.8.4",
-    "unbuild": "0.9.1",
-    "vitest": "0.25.2"
-  }
-}
