--- app-foundations/src/window.ts
@@ -1,7 +1,0 @@
-declare global {
-  interface Window {
-    __previewjs_navigate(componentId: string): void;
-  }
-}
-
-export {};
