--- app-foundations/tailwind.config.js
@@ -1,7 +1,0 @@
-module.exports = {
-  content: ["./src/**/*.{html,ts,tsx}"],
-  theme: {
-    extend: {},
-  },
-  plugins: [],
-};
