--- app-foundations/tsconfig.json
@@ -1,11 +1,0 @@
-{
-  "extends": "../tsconfig.base.json",
-  "compilerOptions": {
-    "lib": ["dom", "dom.iterable"],
-    "module": "ES2020",
-    "moduleResolution": "node",
-    "jsx": "react",
-    "outDir": "dist"
-  },
-  "include": ["src"]
-}
