+++ app/package.json
@@ -35,7 +35,7 @@
     "express": "^4.18.2"
   },
   "devDependencies": {
-    "@previewjs/app-foundations": "workspace:*",
+    "@previewjs/api": "workspace:*",
     "@previewjs/core": "workspace:*",
     "@previewjs/e2e-test-runner": "workspace:*",
     "@previewjs/plugin-react": "workspace:*",
@@ -43,6 +43,7 @@
     "@previewjs/plugin-svelte": "workspace:*",
     "@previewjs/plugin-vue2": "workspace:*",
     "@previewjs/plugin-vue3": "workspace:*",
+    "@previewjs/properties": "workspace:*",
     "@types/express": "4.17.14",
     "@types/fs-extra": "9.0.13",
     "@types/node": "18.11.9",
