+++ pnpm-workspace.yaml
@@ -1,7 +1,6 @@
 packages:
   - "api"
   - "app"
-  - "app-foundations"
   - "chromeless"
   - "config"
   - "core"
