+++ renovate.json
@@ -4,7 +4,6 @@
   "ignoreDeps": [
     "@previewjs/api",
     "@previewjs/app",
-    "@previewjs/app-foundations",
     "@previewjs/chromeless",
     "@previewjs/config",
     "@previewjs/core",
